Convert mostly everything to new webpack modules structure
