+++ packages/core-extensions/src/common/index.ts
@@ -1,11 +1,5 @@
 import { ExtensionWebExports } from "@moonlight-mod/types";
 
-import { react } from "./react";
-import { flux } from "./flux";
-import { stores } from "./stores";
-import { http } from "./http";
-import { fluxDispatcher } from "./fluxDispatcher";
-
 export const webpackModules: ExtensionWebExports["webpackModules"] = {
   components: {
     dependencies: [
@@ -15,9 +9,31 @@ export const webpackModules: ExtensionWebExports["webpackModules"] = {
     ]
   },
 
-  react,
-  flux,
-  stores,
-  http,
-  fluxDispatcher
+  flux: {
+    dependencies: [
+      { ext: "spacepack", id: "spacepack" },
+      "useStateFromStores:function"
+    ]
+  },
+
+  fluxDispatcher: {
+    dependencies: [
+      { ext: "spacepack", id: "spacepack" },
+      "isDispatching",
+      "dispatch"
+    ]
+  },
+
+  react: {
+    dependencies: [
+      { ext: "spacepack", id: "spacepack" },
+      "__SECRET_INTERNALS_DO_NOT_USE_OR_YOU_WILL_BE_FIRED",
+      /\.?version(?:=|:)/,
+      /\.?createElement(?:=|:)/
+    ]
+  },
+
+  stores: {
+    dependencies: [{ ext: "common", id: "flux" }]
+  }
 };
