+++ packages/core-extensions/src/common/webpackModules/flux.ts
@@ -1,0 +1,5 @@
+import spacepack from "@moonlight-mod/wp/spacepack_spacepack";
+
+module.exports = spacepack.findByCode(
+  ["useStateFromStores", ":function"].join("")
+)[0].exports;
