+++ packages/core-extensions/src/common/webpackModules/fluxDispatcher.ts
@@ -1,0 +1,6 @@
+import spacepack from "@moonlight-mod/wp/spacepack_spacepack";
+
+module.exports = spacepack.findByExports(
+  "isDispatching",
+  "dispatch"
+)[0].exports.default;
