+++ packages/core-extensions/src/disableSentry/index.ts
@@ -7,7 +7,8 @@ export const patches: Patch[] = [
     replace: {
       type: PatchReplaceType.Normal,
       match: /default:function\(\){return .}/,
-      replacement: 'default:function(){return require("disableSentry_stub")()}'
+      replacement:
+        'default:function(){return require("disableSentry_stub").proxy()}'
     }
   },
   {
@@ -38,41 +39,5 @@ export const patches: Patch[] = [
 ];
 
 export const webpackModules: ExtensionWebExports["webpackModules"] = {
-  stub: {
-    run: function (module, exports, require) {
-      const logger = moonlight.getLogger("disableSentry");
-
-      const keys = [
-        "setUser",
-        "clearUser",
-        "setTags",
-        "setExtra",
-        "captureException",
-        "captureCrash",
-        "captureMessage",
-        "addBreadcrumb"
-      ];
-
-      module.exports = () =>
-        new Proxy(
-          {},
-          {
-            get(target, prop, receiver) {
-              if (prop === "profiledRootComponent") {
-                return (arg: any) => arg;
-              } else if (prop === "crash") {
-                return () => {
-                  throw Error("crash");
-                };
-              } else if (keys.includes(prop.toString())) {
-                return (...args: any[]) =>
-                  logger.debug(`Sentry calling "${prop.toString()}":`, ...args);
-              } else {
-                return undefined;
-              }
-            }
-          }
-        );
-    }
-  }
+  stub: {}
 };
