+++ packages/core-extensions/src/moonbase/index.tsx
@@ -1,6 +1,5 @@
 import { ExtensionWebExports } from "@moonlight-mod/types";
 import ui from "./ui";
-import { stores } from "./stores";
 import { DownloadIconSVG, TrashIconSVG } from "./types";
 
 export const webpackModules: ExtensionWebExports["webpackModules"] = {
@@ -8,10 +7,7 @@ export const webpackModules: ExtensionWebExports["webpackModules"] = {
     dependencies: [
       { ext: "common", id: "flux" },
       { ext: "common", id: "fluxDispatcher" }
-    ],
-    run: (module, exports, require) => {
-      module.exports = stores(require);
-    }
+    ]
   },
 
   moonbase: {
@@ -26,13 +22,11 @@ export const webpackModules: ExtensionWebExports["webpackModules"] = {
     ],
     entrypoint: true,
     run: (module, exports, require) => {
-      const settings = require("settings_settings");
+      const settings = require("settings_settings").Settings;
       const React = require("common_react");
       const spacepack = require("spacepack_spacepack").spacepack;
       const { MoonbaseSettingsStore } =
-        require("moonbase_stores") as ReturnType<
-          (typeof import("./stores"))["stores"]
-        >;
+        require("moonbase_stores") as typeof import("./webpackModules/stores");
 
       settings.addSection("Moonbase", "Moonbase", ui(require), null, -2, {
         stores: [MoonbaseSettingsStore],
